Add -zero flag
When the -zero flag is given, errcheck will always os.Exit with a status
code of 0.  This option is intended for use with Emacs' flycheck.  If a
checker command exit status != 0 and no messages apply to the current
buffer-file-name, flycheck emits the message:

  Checker %S returned non-zero exit code %s, but no errors from output:
  %s\nChecker definition probably flawed.

This can be common when using errcheck with flycheck, since errcheck
runs against the entire package, rather than a single file.