Fix panic when using line number pragmas.
This commit fixes an issue where generated files that use line pragmas would cause errcheck
to panic. This occurred because errcheck would try to look up the line from the file but
the file did not exist in the checker so an "index out of bounds" panic would occur.

The fix does not try to parse every generated file but instead opts to simply return the
line number and character position. The actual line itself is returned as blank.