Fix internal package directory structure
The package present in the internal directory is named errcheck, but is
not present in a subdirectory. `go get github.com/kisielk/errcheck`
fails on go tip due to this.

This change places the files belonging to the internal errcheck package
in a subdirectory of internal and makes it follow the internal package
convention to be introduced in go 1.5.