Ignore calls that have the comment IGNORE_ERRCHECK
Specifically an ast.CommentGroup associated with the node has to begin
with "IGNORE_ERRCHECK". If such a CommentGroup exists, the call is
ignored. This is useful when you know for a particular call that no
error can possibly be returned, or when you don't care if an error is
returned.