Ignore based on Comment instead of CommentGroup
If any ast.Comment associated with the node begins with
"ERRCHECK_IGNORE" the call is ignored. Previously only ast.CommentGroups
beginning with "ERRCHECK_IGNORE" ignored a call, but that necessitated
always starting every comment group with "ERRCHECK_IGNORE". As an
example, here's a call that would not have been ignored before but is
now ignored:

// Does something.
// IGNORE_ERRCHECK: this cannot return an error because of x.
result, _ = doSomething()