Calculate correct position with comments before the package keyword
Detect the file beginning by using `pass.Fset.File(f.Pos())` instead of
assuming that `f.Pos()` is always the beginning of the file.

`f.Pos()` is not the beginning of the file when there is for instance a
comment before the package, which can lead to wrong reportings of the
actual error position.

Signed-off-by: Roman Mohr <rmohr@google.com>