docs(README.md): Add abspath, mod, and ignoregenerated flag docs
Noticed that some useful flags were not documented in the README.md. I've added the these in what I think are the relevant sections. I did a minimal explanation of the flags, but can expand on them.

I erred on the side of not expanding on -ignoregenerated for example because the regex might change in the future, and documentation tends to be a forgotten update...

Let me know if you need any changes, or anything else I can add to this PR.